Unmake the bridge as a concourse resource
This removes the bits that make the bridge code a concourse resource.
It turns out that we need to be able to pass in the ident-hash value
to the task, which is something that can't be done during a 'get step'
but it can be done as part of a 'task step'.