Fix up resource content types in s3 bucket (temporary)
All the resources just show up as `application/octet-stream` and prompts
the user to download the file when going to the s3 url in the browser.
Set the `Content-Disposition` to `inline` and set the `Content-Type` to
whatever we have in `media_type` in the corresponding `.metadata.json`
file.  Also store everything in `.metadata.json` in the metadata field
together with the s3 file.

This is a quick way to fix the content types but we should probably do
this as part of the s3 concourse resource, probably something we need to
write ourselves.