Concourse pipeline using slack to dump book to s3
Use slack messages to trigger the dump script to download a book to an s3
bucket.

The slack message needs to be in the format:

```
@ce-bot rap-spike get book <book-ident-hash>
```

The variables are in `dump-book-to-s3-vars.yml`.

To create / update this pipeline:

```
fly -t <target> sp -p dump-book-to-s3 -c dump-book-to-s3.yml -l dump-book-to-s3-vars.yml
```